+++ user_docs/en/changes.md
@@ -29,6 +29,8 @@ Unicode CLDR has also been updated.
   * This can be disabled in the "Add-on Store" category of settings.
   * NVDA checks daily for add-on updates.
   * Only updates within the same channel will be checked (e.g. installed beta add-ons will only notify for updates in the beta channel).
+* The volume of the other applications can be adjusted by `NVDA+alt+pageUp` and `NVDA+alt+pageDown`. (#16052, @mltony)
+* The sound of the other applications can be muted with `NVDA+alt+delete`. (#16052, @mltony)
 
 ### Changes